+++ phases/ir_generation/bwsl_ir_analysis.h
@@ -111,21 +111,21 @@ struct IRAnalysis {
 
 // Output slot definitions (for usedOutputMask)
 namespace OutputSlot {
-    constexpr u32 POSITION = 0;       // output.position -> BuiltIn Position
-    constexpr u32 COLOR    = 1;       // output.color -> Location 0 (fragment)
-    constexpr u32 COLOR0   = COLOR;
-    constexpr u32 VARYING0 = 2;       // First varying output
-    constexpr u32 VARYING1 = 3;
-    constexpr u32 VARYING2 = 4;
-    constexpr u32 VARYING3 = 5;
+    constexpr u8 POSITION = 0;       // output.position -> BuiltIn Position
+    constexpr u8 COLOR    = 1;       // output.color -> Location 0 (fragment)
+    constexpr u8 COLOR0   = COLOR;
+    constexpr u8 VARYING0 = 2;       // First varying output
+    constexpr u8 VARYING1 = 3;
+    constexpr u8 VARYING2 = 4;
+    constexpr u8 VARYING3 = 5;
     // Fragment depth
-    constexpr u32 DEPTH    = 16;      // output.depth -> BuiltIn FragDepth
+    constexpr u8 DEPTH    = 16;      // output.depth -> BuiltIn FragDepth
 
-    inline constexpr u32 FragmentColor(u32 location) {
+    inline constexpr u8 FragmentColor(u8 location) {
         return COLOR0 + location;
     }
 
-    inline constexpr u32 FragmentColorLocation(u32 slot) {
+    inline constexpr u8 FragmentColorLocation(u8 slot) {
         return slot - COLOR0;
     }
 }
